Modified Tree Generation
- Fixed bugs in SAXHandler, concurrency failure when reading in
cascading failures of XML
- Updated some method definitions in data classes
- Ignoring (1 - phi) calculations, qMatrix is now filled up completely

+++ src/edu/njit/decaf2/data/State.java
@@ -130,7 +130,7 @@ public class State extends DECAF {
 		return result;
 	}
 	
-	public double getComponentCount(String type){
+	public int getComponentCount(String type){
 		return vector.get(type);
 	}
